requirements.txt
ragas
python-dotenv